Memory leak corrections
All of the changes here were found during testing of the new
interface and are to resolve memory leak issues that were
found at that time.  Very few of them were introduced with the
changes, but the size of the browser objects made them more
obvious and more of a problem.