Fix matching discrepancy and background color
The background color changes I made for the semantic
gathering window caused the color of the background to
be yellow for the part of speech box.  Changed that.

Also noted that WeSay used to match regardless of the case
of the part of speech.  Not quite sure how actually, but
this makes it behave the same.