Display invalid part of speech in red
The old version changed the background behind the list to
red for entries that were in the file but not part of
the configured list.  This matches that change with the
addition that if they select a legitimate entry and then
decide to go back to the original value before they leave
that entry, it is in the drop down list and can be selected.
The color will return to red if they select the original
value.