Initial implementation of GeckoListView
Changes to Gecko Box key handling and tab stops in DetailList

Fixed DetailList so that the tab indexes are in the correct order.
Realized a good bit of the key handling could and needed to
be simplified, in part because upstream objects depend
upon getting key up events.

WeSayListView to GeckoListView conversion

Fixed problems with new entry and delete button on the
Dictionary Control.  Converted list boxes on the Missing
Info Control.  Fixed minor problem in DetailList with
previous memory leak changes.

Add GeckoListBox to the CAWL-SIL word gathering screen

Added the list box to this screen.  Had to make some minor
changes because the data type is not exactly the same
as the other two.  Also had to deal with the fact that due
to some strange code, it is possible for the box to come
up with a height of 0.

Apply maximum length to List Box

Add a maximum length to all list box entries.
Any entries longer than that will be truncated
and an ellipsis added to the value that is
displayed.