WeSayTextBox changes for dual Gecko/WinForms implementation
Added changes so that both GeckoWebBrowser or WinForms
version of WeSayTextBox could be used, controlled by
environment variable WeSayGecko.  If it is defined and
contains "Y" then the new webbrowser version is used.

Complete changes for dual Gecko/WinForm implementation

This completes the changes required to implement the various
controls completed so far to interfaces and to allow the
GeckoBrowser to be used by specifying the WeSayGecko
environment variable. Changes were made throughout to make
sure that the code is identical to the current base release
when running as WinForms.  The sole exception is the
ComboBox, which is being populated somewhat differently
than in the original implementation.